+++ lib/spreadsheet_architect/monkey_patches/axlsx_column_width.rb
@@ -1,28 +1,8 @@
-module SpreadsheetArchitect
-  module AxlsxColumnWidthPatch
-    def initialize(*args)
-      @width = nil
-      super
-    end
-
-    def width=(v)
-      if v.nil?
-        @custom_width = false
-        @width = nil
-      elsif @width.nil? || @width < v+5
-        @custom_width = true
-        @best_fit = true
-        @width = v + 5
-      else
-        @custom_width = true
-        @width = v
-      end
-    end
-  end
-end
-
 module Axlsx
-  class Col
-    prepend ::SpreadsheetArchitect::AxlsxColumnWidthPatch
+  class Cell
+    def string_width(string, font_size)
+      font_scale = font_size / 10.0
+      (string.to_s.size + 3) * font_scale
+    end
   end
 end